Instant change
this allows the map to react to changes to the placemark data if it is updated. i don't know if we want this or if we should keep it as before, when the map was using a private instance of the placemark data

i'm leaning towards not letting the map instantly update because it could fuck some shit up.. maybe there should be a refresher or something. let a nigga know